+++ tests/correctness/framework/OptionOverriding/run.py
@@ -25,6 +25,15 @@ class PySysTest(XpybuildBaseTest):
 		
 		self.assertGrep(file='xpybuild-options.out', expr="Traceback", contains=False)
 		
-		# default options shouldn't contain any objects with no nice string representation
-		self.assertGrep(file='xpybuild-options.out', expr="at 0x", contains=False)
 		
+		self.logFileContents('build-output/BUILD_WORK/targets/MyTarget/implicit-inputs/_OUTPUT_DIR_.defaults.txt.txt', tail=True)
+		
+		# default options shouldn't contain any objects with no nice repr representation, 
+		# least in the implicit inputs file whcih is used to decide rebuilding
+		self.assertGrep(file='build-output/BUILD_WORK/targets/MyTarget/implicit-inputs/_OUTPUT_DIR_.defaults.txt.txt', expr="at 0x", contains=False)
+		# check we did include a wide range of options as a result of the addHashableImplicitInputOption call
+		self.assertGrep(file='build-output/BUILD_WORK/targets/MyTarget/implicit-inputs/_OUTPUT_DIR_.defaults.txt.txt', expr="native.")
+
+		self.assertGrep(file='build-output/BUILD_WORK/targets/MyTarget/implicit-inputs/_OUTPUT_DIR_.defaults.txt.txt', expr="^None$", contains=False) # none items should be filtered  out
+		self.assertGrep(file='build-output/BUILD_WORK/targets/MyTarget/implicit-inputs/_OUTPUT_DIR_.defaults.txt.txt', expr="addHashableImplicitInput str expectedval")
+		self.assertGrep(file='build-output/BUILD_WORK/targets/MyTarget/implicit-inputs/_OUTPUT_DIR_.defaults.txt.txt', expr="testoption.default")